+++ tests/tester/src/test252_parser/edition.rs
@@ -7,7 +7,7 @@ use std::fmt::Display;
 
 use serde_repr::{Deserialize_repr, Serialize_repr};
 
-use crate::read::{MetaData, TestFlag};
+use super::structs::{MetaData, TestFlag};
 
 /// Minimum edition required by a specific feature in the `test262` repository.
 static FEATURE_EDITION: phf::Map<&'static str, SpecEdition> = phf::phf_map! {
@@ -259,7 +259,7 @@ static FEATURE_EDITION: phf::Map<&'static str, SpecEdition> = phf::phf_map! {
     clap::ValueEnum,
 )]
 #[repr(u8)]
-pub(crate) enum SpecEdition {
+pub enum SpecEdition {
     /// ECMAScript 5.1 Edition
     ///
     /// <https://262.ecma-international.org/5.1>
