+++ tests/tester/src/test252_parser/mod.rs
@@ -1,0 +1,8 @@
+mod structs;
+mod edition;
+mod read;
+
+pub use structs::{TestFlags, Ignored, ErrorType, Harness, Outcome, Phase, Test, TestSuite};
+pub use read::{read_suite, read_harness, read_test};
+
+pub use edition::SpecEdition;
